Multiple changes to the project. The big ones are:
1. Removed deprecated source files (geye & geyeview) from gorgoneye.pro
to allow the project to compile with the new Tracker class.

2. Most of the important scaffolding for dynamic detection algorithm
switching is in place. See tracker.h/cpp (context), detector.h/cpp
(strategy) and haar/feature.h/cpp (concreate strategies).

3. Extended each tracker (face, eyes, gaze, etc) by creating an abstract
class (Tracker) in order to ease implementation and hopefully use.
Similar in style to the strategy pattern, but intent different.
Virtualised functions are track() and setDetector().

4. Added work-in-progess MS Visio diagram to show what's going on within
the Tracker and Detector classes and sub-classes.