Changes to Model class and testing UI geyetracker. Changelog includes changes in revision 47 as well.
>> Model class now contains private members for Store and VideoCapture.
These are used internally by the model to grab frames from the webcam
and process them, ideally tracking the user's features and updating the
ROIs and flags available in Store.

>> Model is now instantiated with an integer denoting the OpenCV device
to use for webcam capturing. A Store object and VideoCapture object is
created in the model constructor.

>> The broken revision was caused by a Model in the local scope
overshadowing the class' Model in the private scope.

>> Test UI geyetracker has been modified to reflect the changes to
Model, while retaining dynamic UI generation of tracker/detector
configuration parameters. The test UI/controller successfully
instantiates a Model object which is able to capture frames from the
webcam, process it, as well as obtain the required CV Mat images and
ROIs to display in the QT Painter object.