Summary: CAMShift ported, initial hybrid HAAR/CAMShift for face tracking working, miscellaneous code tweaks
The important stuff:

>> Static CAMShift code successfully ported to FeatureDetector.
Configuration still needs to be made generic.

>> Haar detector has a few code tweaks. Now the CascadeClassifier is a
static variable within locate() instead of being a private member. The
reasoning for this is that a variable should be a Detector member only
if it is required externally. The CAMShift detector also follows this
convention.

>> BaseTracker now has a static BaseDetector pointer, to be shared by
all Trackers when enabling/disabling the Tracker. This should improve
memory usage and code execution slightly.

>> GEyeTracker ROI drawing code has been tweaked to remove the redundant
variable faceLoc.